+++ src/printer.rs
@@ -46,7 +46,10 @@ fn pr_type(t: &Type, print_readably: bool) -> String {
             s.push_str(&format!("\"{}\"", repr));
         }
         Type::Atom(atom) => {
-            s.push_str(&pr_type(&atom.borrow(), print_readably));
+            s.push_str(&format!(
+                "(atom {})",
+                pr_type(&atom.borrow(), print_readably)
+            ));
         }
         Type::List(list) => {
             s.push('(');